[SQL] Split types into a separate crate
