+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/circuit/ICircuit.java
@@ -11,6 +11,9 @@ import javax.annotation.Nullable;
 public interface ICircuit extends IDBSPOuterNode {
     void addOperator(DBSPOperator operator);
     void addDeclaration(DBSPDeclaration declaration);
+    /** Replace an existing declaration with a new one with the same name */
+    void replaceDeclaration(DBSPDeclaration declaration);
+    @Nullable DBSPDeclaration getDeclaration(String name);
     @Nullable
     DBSPViewOperator getView(ProgramIdentifier name);
     Iterable<DBSPOperator> getAllOperators();
