+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/backend/rust/CrateGenerator.java
@@ -3,6 +3,7 @@ package org.dbsp.sqlCompiler.compiler.backend.rust;
 import org.dbsp.sqlCompiler.compiler.DBSPCompiler;
 import org.dbsp.sqlCompiler.compiler.errors.CompilationError;
 import org.dbsp.sqlCompiler.ir.IDBSPNode;
+import org.dbsp.util.Linq;
 import org.dbsp.util.Utilities;
 
 import java.io.File;
@@ -21,15 +22,18 @@ public class CrateGenerator {
     public static final String CARGO = "Cargo.toml";
     public static final String LIB = "lib.rs";
     final List<IDBSPNode> toWrite;
+    final List<CrateGenerator> dependencies;
 
     public CrateGenerator(File baseDirectory, String crateName) {
         this.crateName = crateName;
         this.baseDirectory = baseDirectory;
         this.toWrite = new ArrayList<>();
         this.used = new RustWriter.StructuresUsed();
+        this.dependencies = new ArrayList<>();
     }
 
     void generateCargo(PrintStream stream) {
+        stream.println("# Automatically generated by sql-to-dbsp");
         stream.println("[package]");
         stream.print("name = \"");
         stream.print(this.crateName);
@@ -40,6 +44,16 @@ public class CrateGenerator {
                 edition = "2021"
                 publish = false
                 
+                [dev-dependencies]
+                uuid = { version = "1.6.1" }
+
+                [lib]
+                path = "src/lib.rs"
+                doctest = false
+                
+                [target.'cfg(not(target_env = "msvc"))'.dependencies]
+                tikv-jemallocator = { version = "0.5.4", features = ["profiling", "unprefixed_malloc_on_supported_platforms"] }
+
                 [dependencies]
                 paste = { version = "1.0.12" }
                 derive_more = { version = "0.99.17", features = ["add", "not", "from"] }
@@ -54,29 +68,12 @@ public class CrateGenerator {
                 rust_decimal_macros = { version = "1.36" }
                 serde_json = { version = "1.0.127", features = ["arbitrary_precision"] }
                 rkyv = { version = "0.7.45", default-features = false, features = ["std", "size_64"] }
-                
-                [target.'cfg(not(target_env = "msvc"))'.dependencies]
-                tikv-jemallocator = { version = "0.5.4", features = ["profiling", "unprefixed_malloc_on_supported_platforms"] }
-                
-                [dev-dependencies]
-                uuid = { version = "1.6.1" }
-                
-                [patch.crates-io]
-                datafusion = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
-                datafusion-common = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
-                datafusion-expr = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
-                datafusion-functions = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
-                datafusion-functions-aggregate = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
-                datafusion-physical-expr = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
-                datafusion-physical-plan = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
-                datafusion-proto = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
-                datafusion-sql = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
-                
-                [lib]
-                path = "src/lib.rs"
-                doctest = false""";
+                """;
         cargo = cargo.replace("$ROOT", relativePath);
         stream.println(cargo);
+        for (CrateGenerator dep: this.dependencies) {
+            stream.println(dep.crateName + " = { path = " + Utilities.doubleQuote("../" + dep.crateName) + " }");
+        }
     }
 
     public void add(IDBSPNode node) {
@@ -107,10 +104,16 @@ public class CrateGenerator {
             throw new RuntimeException("Could not create directory " + Utilities.singleQuote(src.getPath()));
         File lib = new File(src, LIB);
         PrintStream rustStream = new PrintStream(Files.newOutputStream(lib.toPath()));
-        RustFileWriter writer = new RustFileWriter(rustStream);
+        RustFileWriter writer = new RustFileWriter(rustStream).withUdf(false);
+        writer.setUsed(this.used);
+        writer.addDependencies(Linq.map(this.dependencies, d -> d.crateName));
         for (IDBSPNode node: this.toWrite)
             writer.add(node);
         writer.writeAndClose(compiler);
+    }
+
+    public void addDependence(CrateGenerator crate) {
+        this.dependencies.add(crate);
     }
 
     public void setUsed(RustWriter.StructuresUsed used) {
