+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/backend/rust/RustCratesWriter.java
@@ -1,6 +1,7 @@
 package org.dbsp.sqlCompiler.compiler.backend.rust;
 
 import org.dbsp.sqlCompiler.circuit.DBSPCircuit;
+import org.dbsp.sqlCompiler.circuit.DBSPDeclaration;
 import org.dbsp.sqlCompiler.compiler.DBSPCompiler;
 import org.dbsp.sqlCompiler.compiler.errors.CompilationError;
 import org.dbsp.sqlCompiler.ir.IDBSPInnerNode;
@@ -11,7 +12,6 @@ import java.io.File;
 import java.io.IOException;
 import java.io.PrintStream;
 import java.nio.file.Files;
-import java.util.List;
 
 /** This class helps generate Rust code.
  * It is given a set of circuit and functions and generates a code in multiple crates. */
@@ -40,24 +40,47 @@ public class RustCratesWriter extends RustWriter {
         File rootDirectory = this.rootDirectory();
         File cargo = new File(rootDirectory, CrateGenerator.CARGO);
         PrintStream cargoStream = new PrintStream(Files.newOutputStream(cargo.toPath()));
+        cargoStream.println("# Automatically generated by sql-to-dbsp");
         cargoStream.println("[workspace]");
         cargoStream.println("members = [");
 
         CrateGenerator main = new CrateGenerator(rootDirectory, "main");
         CrateGenerator types = new CrateGenerator(rootDirectory, "types");
+        main.addDependence(types);
         types.setUsed(used);
+        main.setUsed(new StructuresUsed());
         cargoStream.println("   " + Utilities.doubleQuote(types.crateName) + ",");
         for (IDBSPNode node: this.toWrite) {
             if (node.is(IDBSPInnerNode.class))
                 types.add(node);
-            else
+            else {
+                DBSPCircuit circuit = node.to(DBSPCircuit.class);
                 main.add(node);
+                // Move all declarations to the types crate as inner nodes
+                for (DBSPDeclaration decl: circuit.declarations)
+                    types.add(decl.item);
+                circuit.declarationMap.clear();
+                circuit.declarations.clear();
+            }
         }
         types.write(compiler);
         main.write(compiler);
 
         cargoStream.println("   " + Utilities.doubleQuote(main.crateName));
         cargoStream.println("]");
+
+        cargoStream.println("""
+                [patch.crates-io]
+                datafusion = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
+                datafusion-common = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
+                datafusion-expr = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
+                datafusion-functions = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
+                datafusion-functions-aggregate = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
+                datafusion-physical-expr = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
+                datafusion-physical-plan = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
+                datafusion-proto = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }
+                datafusion-sql = { git = "https://github.com/ryzhyk/datafusion.git", rev = "f561db7" }""");
+
         cargoStream.close();
     }
 }
