+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/backend/rust/RustFileWriter.java
@@ -2,24 +2,24 @@ package org.dbsp.sqlCompiler.compiler.backend.rust;
 
 import org.dbsp.sqlCompiler.circuit.DBSPCircuit;
 import org.dbsp.sqlCompiler.compiler.DBSPCompiler;
-import org.dbsp.sqlCompiler.ir.DBSPFunction;
 import org.dbsp.sqlCompiler.ir.IDBSPInnerNode;
 import org.dbsp.sqlCompiler.ir.IDBSPNode;
 import org.dbsp.util.IndentStream;
 import org.dbsp.util.IndentStreamBuilder;
 import org.dbsp.util.ProgramAndTester;
-import org.dbsp.util.Utilities;
 
 import java.io.IOException;
 import java.io.PrintStream;
 import java.nio.charset.StandardCharsets;
-import java.util.List;
 
 /** This class helps generate Rust code.
  * It is given a set of circuit and functions and generates a compilable Rust file. */
 public class RustFileWriter extends RustWriter {
     final PrintStream outputStream;
     boolean slt = false;
+    boolean generateUdfInclude = true;
+    boolean findUsed = true;
+    StructuresUsed used = new StructuresUsed();
 
     /** Preamble used when generating Rust code. */
     String rustPreamble() {
@@ -42,18 +42,23 @@ public class RustFileWriter extends RustWriter {
         this.slt = true;
         return this;
     }
+    
+    public RustFileWriter withUdf(boolean udf) {
+        this.generateUdfInclude = udf;
+        return this;
+    }
 
     public RustFileWriter(String outputFile)
             throws IOException {
         this(new PrintStream(outputFile, StandardCharsets.UTF_8));
     }
 
-    void generateStructures(StructuresUsed used, IndentStream stream) {
-        super.generateStructures(used, stream);
+    void generateStructures(IndentStream stream) {
+        super.generateStructures(this.used, stream);
         if (this.slt) {
             stream.append("#[cfg(test)]").newline()
                     .append("sltsqlvalue::to_sql_row_impl! {").increase();
-            for (int i : used.tupleSizesUsed) {
+            for (int i : this.used.tupleSizesUsed) {
                 if (i <= 10)
                     // These are already pre-declared
                     continue;
@@ -64,10 +69,10 @@ public class RustFileWriter extends RustWriter {
         }
     }
 
-    String generatePreamble(StructuresUsed used) {
+    String generatePreamble() {
         IndentStream stream = new IndentStreamBuilder();
         stream.append(commonPreamble);
-        long max = used.getMaxTupleSize();
+        long max = this.used.getMaxTupleSize();
         if (max > 120) {
             // this is just a guess
             stream.append("#![recursion_limit = \"")
@@ -77,26 +82,12 @@ public class RustFileWriter extends RustWriter {
         }
 
         stream.append("""
-            #[cfg(test)]
-            use hashing::*;""")
+                      #[cfg(test)]
+                      use hashing::*;""")
                 .newline();
         stream.append(this.rustPreamble())
                 .newline();
-        this.generateStructures(used, stream);
-
-        String stubs = Utilities.getBaseName(DBSPCompiler.STUBS_FILE_NAME);
-        stream.append("mod ")
-                .append(stubs)
-                .append(";")
-                .newline()
-                .append("mod ")
-                .append(Utilities.getBaseName(DBSPCompiler.UDF_FILE_NAME))
-                .append(";")
-                .newline()
-                .append("use crate::")
-                .append(stubs)
-                .append("::*;")
-                .newline();
+        this.generateStructures(stream);
         return stream.toString();
     }
 
@@ -110,9 +101,19 @@ public class RustFileWriter extends RustWriter {
         this.toWrite.add(node);
     }
 
+    public void setUsed(StructuresUsed used) {
+        this.used = used;
+        this.findUsed = false;
+    }
+
     public void write(DBSPCompiler compiler) {
-        StructuresUsed used = this.analyze(compiler);
-        this.outputStream.println(generatePreamble(used));
+        if (this.findUsed)
+            this.used = this.analyze(compiler);
+        this.outputStream.println(generatePreamble());
+        if (this.generateUdfInclude)
+            this.outputStream.println(generateUdfInclude());
+        for (String dep: this.dependencies)
+            this.outputStream.println("use " + dep + "::*;");
         for (IDBSPNode node: this.toWrite) {
             String str;
             IDBSPInnerNode inner = node.as(IDBSPInnerNode.class);
